+++ projects/swimlane/ngx-ui/src/lib/components/date-time/date-time.component.ts
@@ -239,7 +239,7 @@ export class DateTimeComponent implements OnDestroy, ControlValueAccessor {
   dialog: any;
   dialogModel: moment.Moment;
   hour: number;
-  minute: number;
+  minute: string;
   amPmVal: string;
   displayValue = '';
   modes = ['millisecond', 'second', 'minute', 'hour', 'date', 'month', 'year'];
@@ -293,20 +293,20 @@ export class DateTimeComponent implements OnDestroy, ControlValueAccessor {
   setDialogDate(date) {
     this.dialogModel = this.createMoment(date);
     this.hour = +this.dialogModel.format('hh');
-    this.minute = +this.dialogModel.format('mm');
+    this.minute = this.dialogModel.format('mm');
     this.amPmVal = this.dialogModel.format('A');
   }
 
   dateSelected(date: any): void {
     this.dialogModel = this.createMoment(date);
     this.hour = +this.dialogModel.format('hh');
-    this.minute = +this.dialogModel.format('mm');
+    this.minute = this.dialogModel.format('mm');
     this.amPmVal = this.dialogModel.format('A');
   }
 
   minuteChanged(newVal: number): void {
     this.dialogModel = this.dialogModel.clone().minute(newVal);
-    this.minute = +this.dialogModel.format('mm');
+    this.minute = this.dialogModel.format('mm');
   }
 
   hourChanged(newVal: number): void {